extract more utils
